Create Discretisation with Decision Trees
Discretisation with Decision Trees consists in using a decision tree to identify the optimal bins. 
When a decision tree makes a decision, it assigns an observation to one of n end leaves. 
Therefore, any decision tree will generate a discrete output, which values are the predictions 
at each of its n leaves.